% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMaire2019DistanceMatrix}
\alias{exampleMaire2019DistanceMatrix}
\title{Maire 2019 distance matrix}
\format{
An Rdata file
}
\usage{
exampleMaire2019DistanceMatrix
}
\value{
A distance matrix
}
\description{
Distance matrix from Maire et al. (2019) for the example
multilevel/multivariate meta-analysis. See dat.maire2019 in the metadat
package.
}
\keyword{datasets}
