% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleRobmaLui2015}
\alias{exampleRobmaLui2015}
\title{Example RoBMA fit to Lui 2015 data}
\format{
An RData file
}
\usage{
exampleRobmaLui2015
}
\value{
An object of class RoBMA.
}
\description{
RoBMA fit (with removed posterior samples) for
the example analysis on the Lui 2015 data example from AMPPS
(vignette at https://cran.r-project.org/web/packages/RoBMA/vignettes/Tutorial.html)
}
\keyword{datasets}
